Move bignumber utils to hardhat/common
