+++ packages/hardhat-core/package.json
@@ -54,6 +54,7 @@
     "console.sol"
   ],
   "devDependencies": {
+    "@types/bn.js": "^5.1.0",
     "@nomiclabs/eslint-plugin-hardhat-internal-rules": "^1.0.0",
     "@types/chai": "^4.2.0",
     "@types/chai-as-promised": "^7.1.3",
@@ -74,6 +75,8 @@
     "@types/ws": "^7.2.1",
     "@typescript-eslint/eslint-plugin": "4.29.2",
     "@typescript-eslint/parser": "4.29.2",
+    "bignumber.js": "^9.0.2",
+    "bn.js": "^5.1.0",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "eslint": "^7.29.0",
